requirements.txt
Flask==2.2.2
mirai==0.1.5
PyYAML==6.0
